kubernetes: drop v1.15.3 recipe
Issue: LINUXEXEC-7608

Removed kubernetes v1.15.3 recipe

Removed go v1.12 recipes required for kubernetes v1.15.3.
v1.17.x requires go version 1.13.4 or greater.
go in oe-core is 1.14.1 which is also the internal version.

Removed patches that were backported to v1.15.3, which are
now merged and present in v1.17.1.
	- kubernetes/0001-Hide-bearer-token-in-logs.patch
	- kubernetes/0001-fix-compiling-failure-execvp-bin-bash-Argument-list-.patch
	- kubernetes/0001-version.sh-get-rid-of-dirty.patch

In essence, this is sync to upstream meta-virutalization.

(LOCAL REV: NOT UPSTREAM) - wrlinux specific

Signed-off-by: Sakib Sajal <sakib.sajal@windriver.com>