kubernetes: uprev v1.17.1 -> v1.17.5
Issue: LINUXEXEC-7608

Changed SRCREV to point to v1.17.5

This recipe is derived from the kubernetes_git.bb recipe.
We leave the _git.bb recipe intact and have this as
WR specific as upstream has opted for an update
cadence which does not match our product requirements.

(LOCAL REV: NOT UPSTREAM) - wrlinux specific

Signed-off-by: Sakib Sajal <sakib.sajal@windriver.com>