kubernetes: use specific directory for stable recipe
Issue: LINUXEXEC-10891

Currently our stable version share the same direcotry for patches with
upstream development version. This is not good as when upstream changes
a patch, we are likely to meet do_patch error.

So we use our own directory to hold patches.

(LOCAL REV: NOT UPSTREAM) -- WRLinux Specific Patch

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>