docker: Add kernel-module-xt-nat to RRECOMMENDS
Issue: LINCD-7845

Fixed when kernel-module-xt-nat is not installed:
$ docker run --rm -it -p 80:80 alpine

docker: Error response from daemon: driver failed programming external connectivity on endpoint elated_cori
Try `iptables -h' or 'iptables --help' for more information.
 (exit status 2)).

(LOCAL REV: NOT UPSTREAM) - Sent to upstream on 20220119, but upstream is
looking for a way to fix all of the rrecommends issues, which has not been done yet.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>