containers: introduce container-host class
Issue: LINUXEXEC-25377

Introducing a small (at the moment) class that represents configuration
and processing required to prepare a target image to be a container
host.

A recipe that requires container configuration should inherit this
class, and the container-host-config package will be added as a
RDEPENDS, and install common configuration files.

In the future, additional functionality or dependencies will be added
here to synchronize the configuration of multiple container host
packages.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>

(LOCAL REV: NOT UPSTREAM) -- sent to meta-virt on 20230302

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>