containers: add container-host-config recipe
Issue: LINUXEXEC-25377

This is a configuration only recipe that produces a package which
installs some common configuration files.

In this introduction we have both registries.conf and storage.conf.

Packages that require these files should RDEPEND on this package
(or inherit container-host.bbclass) and the files will be installed.

If conflicting requirements for these global configuration files
arise, they can be resolved through additions to this recipe, or by
providing a higher priority version of the .conf files.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>

(LOCAL REV: NOT UPSTREAM) -- sent to meta-virt on 20230302

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>