skopeo: use container-host bbclass to provide configuration
Issue: LINUXEXEC-25377

Instead of providing storage and registries configuration files
in this package, we inherit container-host which will provide a
common definition of these configs.

This allows multiple packages to ensure that the configuration
files are present, and not conflict in their installation.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>

(LOCAL REV: NOT UPSTREAM) -- sent to meta-virt on 20230302

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>