container-host-config: provide /etc/containers/policy.json
Issue: LINUXEXEC-25377

The /etc/containers/policy.json[1] file is used to specify verification
policy. For now, we can see it's used by both cri-o and skopeo. To avoid
conflict, we use container-host-config to provide this file and make both
skopeo and cri-o depend on it.

[1] https://github.com/containers/image/blob/main/docs/containers-policy.json.5.md

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>

(LOCAL REV: NOT UPSTREAM) -- sent to meta-virt on 20230302

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>